[refactor] likes 관련 리포지토리 삭제, course 리포지토리로 이동
- @Operation 및 @SwaggerErrorApi 추가
