+++ src/main/java/com/example/shimpyo/domain/course/controller/LikesController.java
@@ -1,6 +1,9 @@
-package com.example.shimpyo.domain.likes.controller;
+package com.example.shimpyo.domain.course.controller;
 
-import com.example.shimpyo.domain.likes.service.LikesService;
+import com.example.shimpyo.domain.course.service.LikesService;
+import com.example.shimpyo.global.SwaggerErrorApi;
+import com.example.shimpyo.global.exceptionType.MemberExceptionType;
+import com.example.shimpyo.global.exceptionType.TouristException;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -13,12 +16,14 @@ import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/likes")
+@RequestMapping("/api/course")
 public class LikesController {
     private final LikesService likesService;
 
-    @PatchMapping
     @Operation(summary = "관광지 찜 추가/삭제")
+    @SwaggerErrorApi(type = {TouristException.class, MemberExceptionType.class},
+            codes = {"TOURIST_NOT_FOUND", "MEMBER_NOT_FOUND"})
+    @PatchMapping
     public ResponseEntity<Void> toggleLikeTourist(@RequestBody Map<String, Long> requestDto) {
         likesService.toggleLikeTourist(requestDto.get("id"));
         return ResponseEntity.ok().build();
