+++ src/main/java/com/example/shimpyo/domain/user/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.example.shimpyo.domain.user.controller;
 
-import com.example.shimpyo.domain.likes.service.LikesService;
+import com.example.shimpyo.domain.course.service.LikesService;
 import com.example.shimpyo.domain.user.dto.TouristLikesResponseDto;
 import com.example.shimpyo.domain.user.service.UserService;
 import com.example.shimpyo.global.BaseException;
@@ -50,6 +50,7 @@ public class UserController {
     }
 
     @Operation(summary = "찜한 관광지 목록")
+    @SwaggerErrorApi(type = {MemberExceptionType.class}, codes = {"MEMBER_NOT_FOUND"})
     @GetMapping("/likes")
     public ResponseEntity<List<TouristLikesResponseDto>> getTouristLikes(@RequestParam("category") String category,
                                                                          @RequestParam("likesId") Long id) {
